+++ src/app/Http/Livewire/Forms/FormCheckbox.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Livewire;
+namespace App\Http\Livewire\Forms;
 
-use App\Http\Livewire\Form;
+use App\Http\Livewire\forms\Form;
 
 class FormCheckbox extends Form
 {
