+++ src/resources/views/fields/edit.blade.php
@@ -19,7 +19,7 @@
                         <!--  TITLE SECTION  -->
                         <div class="px-4 py-5 bg-white sm:p-6">
                             @php($form_elem = "title")
-                            @livewire("form-text", [
+                            @livewire("forms.form-text", [
                                 'form_elem' => $form_elem,
                                 'title' => "Field's title",
                                 'placeHolder' => "Example : Full Name",
@@ -32,7 +32,7 @@
                         <div class="px-4 py-5 bg-white sm:p-6">
 
                             @php($form_elem = "placeholder")
-                            @livewire("form-text", [
+                            @livewire("forms.form-text", [
                                 'form_elem' => $form_elem,
                                 'title' => "Field's placeholder",
                                 'placeHolder' => "The placehold is shown as an example when the field is asked (just like this)",
@@ -99,7 +99,7 @@
 
                         <div class="px-4 py-5 bg-white sm:p-6">
                             @php($form_elem = "descriptive_value")
-                            @livewire("form-checkbox", [
+                            @livewire("forms.form-checkbox", [
                                 'form_elem' => $form_elem,
                                 'title' => "Is that a descriptive value ?",
                                 'hint' => "If checked, it will be displayed in the Persons section.",
@@ -112,7 +112,7 @@
                         <div class="px-4 py-5 bg-white sm:p-6">
 
                             @php($form_elem = "best_descriptive_value")
-                            @livewire("form-checkbox", [
+                            @livewire("forms.form-checkbox", [
                                 'form_elem' => $form_elem,
                                 'title' => "Is that a descriptive value ?",
                                 'hint' => "If checked, it will be displayed in the Persons section.",
